requirements.txt
PyMySQL==0.9.3
Django==1.11.10
django-simple-captcha
